Pre-commit fixes

+++ holoocean_bridge/holoocean_bridge/utils/seatrac_enums.py
@@ -57,7 +57,7 @@ def clamp_int16(v: float) -> int:
     :param v: Value to clamp.
     :return: The value rounded and clamped to [-32768, 32767].
     """
-    return max(-32768, min(32767, int(round(v))))
+    return max(-32768, min(32767, round(v)))
 
 
 def clamp_uint16(v: float) -> int:
@@ -67,4 +67,4 @@ def clamp_uint16(v: float) -> int:
     :param v: Value to clamp.
     :return: The value rounded and clamped to [0, 65535].
     """
-    return max(0, min(65535, int(round(v))))
+    return max(0, min(65535, round(v)))
